refactor: intersection observer target을 한 개로 변경
