+++ src/features/users/routes/Profile/TabMenu/ReviewList.tsx
@@ -1,36 +1,11 @@
-import {
-  FetchNextPageOptions,
-  InfiniteQueryObserverResult,
-} from "@tanstack/react-query";
-import { useRef } from "react";
-
 import ReviewCard from "@/features/reviews/components/ReviewCard";
-import useIntersectionObserver from "@/hooks/useIntersectionObserver";
-
-import { Target } from "./ReviewList.style";
 
 interface ReviewListProps {
   isMine: boolean;
   list: Review[];
-  fetchNextPage: (
-    options?: FetchNextPageOptions | undefined,
-  ) => Promise<InfiniteQueryObserverResult<Review, unknown>>;
-  hasNextPage: boolean | undefined;
 }
 
-export default function ReviewList({
-  isMine,
-  list,
-  fetchNextPage,
-  hasNextPage,
-}: ReviewListProps) {
-  const targetRef = useRef(null);
-
-  useIntersectionObserver({
-    target: targetRef,
-    onIntersect: fetchNextPage,
-    enabled: hasNextPage,
-  });
+export default function ReviewList({ isMine, list }: ReviewListProps) {
   return (
     <>
       {list.map((review) => (
@@ -50,8 +25,6 @@ export default function ReviewList({
           />
         </ReviewCard>
       ))}
-
-      <Target ref={targetRef} />
     </>
   );
 }
